Implemented single active review slider and added clearer comments for understanding
