+++ 07-slider/setup/src/index.css
@@ -56,8 +56,8 @@ Global Styles
   box-sizing: border-box;
 }
 body {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
+    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
   background: var(--clr-grey-10);
   color: var(--clr-grey-1);
   line-height: 1.5;
@@ -156,13 +156,14 @@ Slider
   margin: 0 auto;
   margin-top: 4rem;
   width: 80vw;
-  /* have to have a height */
+  /* must to have a height because of the position of absolute used later */
   height: 450px;
   max-width: 800px;
   text-align: center;
   position: relative;
   display: flex;
-  /* overflow: hidden; */
+  /* was commented */
+  overflow: hidden;
 }
 .person-img {
   border-radius: 50%;
@@ -239,10 +240,13 @@ article {
   left: 0;
   width: 100%;
   height: 100%;
-  /* opacity: 0; */
+  /* was commented */
+  opacity: 0;
   transition: var(--transition);
 }
-/* article.activeSlide {
+
+/* was commented */
+article.activeSlide {
   opacity: 1;
   transform: translateX(0);
 }
@@ -251,4 +255,4 @@ article.lastSlide {
 }
 article.nextSlide {
   transform: translateX(100%);
-} */
+}
